+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 2. SQL 支援類別的成員是 Nullable Enum 資料類型的 property  (資料會嘗試轉換到 Enum，如果失敗則會容錯成 Enum 的預設值，不會變成 null)
 3. `Cast.To` 如果嘗試轉換成 Enum 類別，則會自動容錯成呼叫 Enum.To 函數
 4. Config 新增 `Config.Refresh` 方法，如果應用程式運行過程中可能會改變 `.config` 檔案，可以自行決定何時刷新 Config 中的快取 (cache)
-
+5. 新增 `JXPath.GetValue` Utility 類別方法，允許指定 XPath 來取出特定 `dynamic` 或 JSON 字串中指定路徑位置的數值
 
 ## `2.0.11`